Introduce support for the new "i4i.metal" AWS host instance type
Due to a deficiency in the GO SDK the string value  "I4I_METAL" is sent on the wire, which is not recognised by the
VMC service as a valid host instance type. Thus a bit of hacking is required until the issue with the
GO SDK is fixed.

Additionally, removed references to the R5 host instance type from the documentation, due to a shortage of these
hosts in AWS. The code that concerns them is left to enable users to manage their already deployed instances.

Testing done:
Successfully deployed a single-host i4i.metal SDDC
Successfully destroyed a single-host i4i.metal SDDC

Successfully deployed a single-host i3i.metal SDDC
Successfully destroyed a single-host i3i.metal SDDC

golangci-lint run - passed

Signed-off-by: Dimitar Proynov <proynovd@vmware.com>