Changing the lock position in resourceSrmNodeDelete
TestAccResourceVmcMultipleSrmNodesZerocloud was failing with

 testing_new.go:82: Error running post-test destroy, there may be dangling resources: exit status 1

        Error: task failed: failed to delete SRM node:
com.vmware.skyscraper.common.BadRequestException: Failed to lock site
recovery for sddc id ########

The lock in resourceSrmNodeDelete was after tha actual delete call.
Changed the position of the lock to precede the delete call.

Managed to reproduce the failure of TestAccResourceVmcMultipleSrmNodesZerocloud
before the change

Testing done:

Ran several times  TestAccResourceVmcMultipleSrmNodesZerocloud locally
with success wfter the change

Signed-off-by: Vasil Atanasov <vasila@vmware.com>